build: Push storage container on PRs
This commit adds the ability to push container images to GHCR during the
storage container build while running on pull requests. Useful for
testing container images resulting from PRs.

This also ensures we are separating out the build and publish steps, and
running UT in between, and *not* publishing if the UT fails.

Also. remove cosign.

Closes #161

Signed-off-by: Kyle Mestery <mestery@mestery.com>