storage: ARCH=skylake to avoid using too new ISA
'native' arch will cause gcc to optimize using the latest available
instructions on the build CPU.  If the build CPU is newer than the
runtime CPU, it will fail with "Illegal instruction".  'skylake' is
new enough to support the minimum ISA required by spdk/dpdk but old
enough to be largely available.

Signed-off-by: Boris Glimcher <Boris.Glimcher@emc.com>
Signed-off-by: Steven Royer <sroyer@redhat.com>