Avoid using global variables for NVMe parameters.
In this change, NVMe related global variables are moved into
frontend Server structure. It is good design practise to
avoid global variables in the code. It makes design flexible
and testable. NVMe parameters are made exported from package
to provide possibility for clients to leverage them in xPU
specific code.

Signed-off-by: Artsiom Koltun <artsiom.koltun@intel.com>